Split cli-configuration tests by action
Follow the file naming convention {sut}.{action}.cs:
- cli-configuration.set-command-path.cs (5 tests)
- cli-configuration.clear-command-path.cs (1 test)
- cli-configuration.reset.cs (1 test)
- cli-configuration.all-command-paths.cs (1 test)

Each file has:
- Namespace: CliConfiguration_
- Class: {Action}_Given_ (e.g., SetCommandPath_Given_)
- #region Purpose and #region Design blocks
